// src/types/pulse.ts
export interface PulseToken {
    id: string;
    name: string;
    img: string;
  
    symbol: string;
    addressShort: string;
    addressFull: string;
  
    age: string;
    holders: number;
    kols: number;
    vip: string;
  
    chef: string;
    userSetting: string;
    views: number;
  
    marketCap: string;
    volume: string;
    solFees: string;
  
    txCount: number;
    sniperPercent: number;
    bundlePercent: number;
  }